Refactor function generation
