+++ cs-bindgen-shared/src/ret.rs
@@ -8,7 +8,7 @@ use syn::{spanned::Spanned, Type};
 /// `Primitive` variant. This allows us to specifically identify primitive types
 /// that can be passed across the FFI boundary without additional marshalling (or at
 /// least without the complexity of fully describing the type).
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum ReturnType {
     Default,
     SelfType,
